+++ src/main/java/com/aetherteam/aether/entity/AetherEntityTypes.java
@@ -74,13 +74,13 @@ public final class AetherEntityTypes {
             key -> EntityType.Builder.of(PassiveWhirlwind::new, AetherMobCategory.AETHER_SURFACE_MONSTER).fireImmune().sized(0.6F, 0.8F).clientTrackingRange(8).build(key));
 
     public static final EntityType<EvilWhirlwind> EVIL_WHIRLWIND = register("evil_whirlwind",
-            EntityType.Builder.of(EvilWhirlwind::new, AetherMobCategory.AETHER_SURFACE_MONSTER).fireImmune().sized(0.6F, 0.8F).clientTrackingRange(8).build(key("evil_whirlwind")));
+            key -> EntityType.Builder.of(EvilWhirlwind::new, AetherMobCategory.AETHER_SURFACE_MONSTER).fireImmune().sized(0.6F, 0.8F).clientTrackingRange(8).build(key));
 
     public static final EntityType<AechorPlant> AECHOR_PLANT = register("aechor_plant",
-            EntityType.Builder.of(AechorPlant::new, AetherMobCategory.AETHER_SURFACE_MONSTER).sized(1.0F, 1.0F).clientTrackingRange(8).build(key("aechor_plant")));
+            key -> EntityType.Builder.of(AechorPlant::new, AetherMobCategory.AETHER_SURFACE_MONSTER).sized(1.0F, 1.0F).clientTrackingRange(8).build(key));
 
     public static final EntityType<Cockatrice> COCKATRICE = register("cockatrice",
-            EntityType.Builder.of(Cockatrice::new, AetherMobCategory.AETHER_DARKNESS_MONSTER).sized(0.9F, 2.15F).clientTrackingRange(10).build(key("cockatrice")));
+            key -> EntityType.Builder.of(Cockatrice::new, AetherMobCategory.AETHER_DARKNESS_MONSTER).sized(0.9F, 2.15F).clientTrackingRange(10).build(key));
 
     public static final EntityType<Zephyr> ZEPHYR = register("zephyr",
             EntityType.Builder.of(Zephyr::new, AetherMobCategory.AETHER_SKY_MONSTER).sized(4.5F, 3.5F).clientTrackingRange(10).build(key("zephyr")));